Added consul helper.
