+++ src/app.ts
@@ -1,9 +1,9 @@
 import express from 'express';
-import routers from './routers';
-import config from './config';
 import log4js, { Configuration } from 'log4js';
 import mongoose, { ConnectOptions } from 'mongoose';
-import Consul, { ConsulOptions } from 'consul';
+import config from './config';
+import { getConsulValue } from './helper/consul';
+import routers from './routers';
 
 type EnvType = 'dev' | 'prod';
 
@@ -11,19 +11,6 @@ let env: EnvType = 'prod';
 if (String(process.env.NODE_ENV).trim() === 'dev') {
   env = 'dev';
 }
-
-const consulServer = new Consul(config.consul.server[env] as ConsulOptions);
-
-const prefix = `config/${config.consul.service.name}`;
-
-type ConsulResult = {
-	Value: string | number,
-};
-
-const getConsulValue = async (key: string) => {
-  const result: ConsulResult = await consulServer.kv.get(`${prefix}/${key}`);
-  return result?.Value;
-};
 
 export default async () => {
   const app = express();
